GRAF-85 Auth Guards & safeties + test fixes

+++ src/graffitiphoto/graffitiphoto.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
 import sharp from 'sharp';
 import { MetadataServiceJS } from '../metadata/metadata.servicejs';
 import { PrismaService } from '../prisma/prisma.service';
@@ -6,7 +6,10 @@ import { S3Service } from '../s3/S3service';
 import { LikesEntry } from './dto/request/LikesEntry';
 import { CreateGraffitiPhotoDto } from './dto/request/create-graffitiphoto.dto';
 import { UpdateGraffitiPhotoDto } from './dto/request/update-graffitiphoto.dto';
+
 import { Multer } from 'multer';
+import { AuthService } from '../auth/auth.service';
+import { Request } from 'express';
 
 type File = Express.Multer.File;
 
@@ -14,14 +17,22 @@ type File = Express.Multer.File;
 export class GraffitiPhotoService {
 	constructor(
 		private prisma: PrismaService,
-		@Inject(S3Service)
 		private S3Service: S3Service,
-
-		@Inject(MetadataServiceJS)
 		private MetadataService: MetadataServiceJS,
+		private authService: AuthService,
 	) {}
 
-	async create(createGraffitiPhotoDto: CreateGraffitiPhotoDto, file: File) {
+	async create(
+		createGraffitiPhotoDto: CreateGraffitiPhotoDto,
+		file: File,
+		request: Request,
+	) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
+
 		file.buffer = await this.MetadataService.removeMetadata(file);
 		let md = await sharp(file.buffer).metadata();
 		console.log('metadata111:', md);
@@ -85,7 +96,26 @@ export class GraffitiPhotoService {
 		return entity;
 	}
 
-	async update(id: number, updateGraffitiPhotoDto: UpdateGraffitiPhotoDto) {
+	async update(
+		id: number,
+		updateGraffitiPhotoDto: UpdateGraffitiPhotoDto,
+		request: Request,
+	) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
+
+		let user = await this.authService.getUserFromRequest(request);
+		let entity = await this.prisma.graffitiPhoto.findUniqueOrThrow({
+			where: {
+				id: id,
+			},
+		});
+		if (entity.userId !== user?.userId) {
+			throw new UnauthorizedException('User is not authorized');
+		}
+
 		return await this.prisma.graffitiPhoto.update({
 			where: {
 				id: id,
@@ -94,39 +124,73 @@ export class GraffitiPhotoService {
 		});
 	}
 
-	async addLikedPhoto(id: number, request: LikesEntry) {
-		let entity = await this.prisma.graffitiPhoto.update({
+	async addLikedPhoto(id: number, request: Request) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
+		let user = await this.authService.getUserFromRequest(request);
+		if (!user || !user.userId) {
+			throw new UnauthorizedException('User does not exist');
+		}
+
+		return await this.prisma.graffitiPhoto.update({
 			where: {
 				id: id,
 			},
 			data: {
 				likes: {
-					create: request.userId.map((userId) => ({
-						userId: userId,
-					})),
+					create: {
+						userId: user.userId,
+					},
 				},
 			},
 		});
-		return entity;
 	}
 
-	async removeLikedPhoto(id: number, request: LikesEntry) {
-		let entity = await this.prisma.graffitiPhoto.update({
+	async removeLikedPhoto(id: number, request: Request) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
+		let user = await this.authService.getUserFromRequest(request);
+		if (!user || !user.userId) {
+			throw new UnauthorizedException('User does not exist');
+		}
+
+		return await this.prisma.graffitiPhoto.update({
 			where: {
 				id: id,
 			},
 			data: {
 				likes: {
-					deleteMany: request.userId.map((userId) => ({
-						userId: userId,
-					})),
+					delete: {
+						userId_graffitiPhotoId: {
+							userId: user.userId,
+							graffitiPhotoId: id,
+						},
+					},
 				},
 			},
 		});
-		return entity;
 	}
 
-	async delete(id: number) {
+	async delete(id: number, request: Request) {
+		let isUserLoggedIn = await this.authService.isLoggedIn(request);
+		if (!isUserLoggedIn) {
+			throw new UnauthorizedException('User is not logged in');
+		}
+
+		let user = await this.authService.getUserFromRequest(request);
+		let entity = await this.prisma.graffitiPhoto.findUniqueOrThrow({
+			where: {
+				id: id,
+			},
+		});
+		if (entity.userId !== user?.userId) {
+			throw new UnauthorizedException('User is not authorized');
+		}
+
 		return await this.prisma.graffitiPhoto.delete({
 			where: {
 				id: id,
